Move definitions and definition generators in different packages
