+++ src/org/emoflon/flight/model/generator/ModelGenerator.java
@@ -1,7 +1,9 @@
-package org.emoflon.flight.model.util;
+package org.emoflon.flight.model.generator;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import org.emoflon.flight.model.util.ModelParser;
 
 import Flights.Airport;
 import Flights.Airports;
@@ -99,7 +101,7 @@ public class ModelGenerator {
 		return null;
 	}
 	/**
-	 * @param fileName of the corresponding '.airports' file in '/Flights/src/org/emoflon/flight/model/examples'
+	 * @param fileName of the corresponding '.airports' file in '/Flights/src/org/emoflon/flight/model/definitions'
 	 * @return an airport container filled with airport parsed from the '.airports' file
 	 */
 	protected Airports parseAirportsWithGates(String fileName) {
@@ -109,7 +111,7 @@ public class ModelGenerator {
 		return airports;
 	}
 	/**
-	 * @param fileName of the corresponding '.airports' file in '/Flights/src/org/emoflon/flight/model/examples'
+	 * @param fileName of the corresponding '.airports' file in '/Flights/src/org/emoflon/flight/model/definitions'
 	 * @return a list of airports filled with airport objects parsed from the '.airports' file
 	 */
 	protected List<Airport> parseAirprortListWithGates(String fileName) {
@@ -125,7 +127,7 @@ public class ModelGenerator {
 		return airportList;
 	}
 	/**
-	 * @param fileName of the corresponding '.routes' file in '/Flights/src/org/emoflon/flight/model/examples'
+	 * @param fileName of the corresponding '.routes' file in '/Flights/src/org/emoflon/flight/model/definitions'
 	 * @param airports containing all airports named in the '.routes' file
 	 * @return a route container filled with route objects parsed from the '.routes' file
 	 */
@@ -136,7 +138,7 @@ public class ModelGenerator {
 		return routes;
 	}
 	/**
-	 * @param fileName of the corresponding '.routes' file in '/Flights/src/org/emoflon/flight/model/examples'
+	 * @param fileName of the corresponding '.routes' file in '/Flights/src/org/emoflon/flight/model/definitions'
 	 * @param airports containing all airports named in the '.routes' file
 	 * @return a route list filled with route objects parsed from the '.routes' file
 	 */
@@ -156,7 +158,7 @@ public class ModelGenerator {
 	}
 	
 	/**
-	 * @param fileName of the corresponding '.flights' file in '/Flights/src/org/emoflon/flight/model/examples'
+	 * @param fileName of the corresponding '.flights' file in '/Flights/src/org/emoflon/flight/model/definitions'
 	 * @param airports containing all airports named in the '.flights' file
 	 * @param routes containing all routes named in the '.flights' file
 	 * @param planes containing all planes named in the '.flights' file
@@ -169,7 +171,7 @@ public class ModelGenerator {
 		return flights;
 	}
 	/**
-	 * @param fileName of the corresponding '.flights' file in '/Flights/src/org/emoflon/flight/model/examples'
+	 * @param fileName of the corresponding '.flights' file in '/Flights/src/org/emoflon/flight/model/definitions'
 	 * @param airports containing all airports named in the '.flights' file
 	 * @param routes containing all routes named in the '.flights' file
 	 * @param planes containing all planes named in the '.flights' file
@@ -195,7 +197,7 @@ public class ModelGenerator {
 		return flights;
 	}
 	/**
-	 * @param fileName of the corresponding '.planes' file in '/Flights/src/org/emoflon/flight/model/examples'
+	 * @param fileName of the corresponding '.planes' file in '/Flights/src/org/emoflon/flight/model/definitions'
 	 * @return a plane container filled with plane objects parsed from the '.planes' file
 	 */
 	protected Planes parsePlanes(String fileName) {
@@ -206,7 +208,7 @@ public class ModelGenerator {
 		return planes;
 	}
 	/**
-	 * @param fileName of the corresponding '.planes' file in '/Flights/src/org/emoflon/flight/model/examples'
+	 * @param fileName of the corresponding '.planes' file in '/Flights/src/org/emoflon/flight/model/definitions'
 	 * @return a plane list filled with plane objects parsed from the '.planes' file
 	 */
 	protected List<Plane> parsePlaneList(String fileName) {
@@ -221,7 +223,7 @@ public class ModelGenerator {
 		return planeList;
 	}
 	/**
-	 * @param fileName of the corresponding '.persons' file in '/Flights/src/org/emoflon/flight/model/examples'
+	 * @param fileName of the corresponding '.persons' file in '/Flights/src/org/emoflon/flight/model/definitions'
 	 * @return a person container filled with person objects parsed from the '.persons' file
 	 */
 	protected Persons parsePersons(String fileName) {
@@ -231,7 +233,7 @@ public class ModelGenerator {
 		return persons;
 	}
 	/**
-	 * @param fileName of the corresponding '.persons' file in '/Flights/src/org/emoflon/flight/model/examples'
+	 * @param fileName of the corresponding '.persons' file in '/Flights/src/org/emoflon/flight/model/definitions'
 	 * @return a person list filled with person objects parsed from the '.persons' file
 	 */
 	protected List<Person> parsePersonList(String fileName) {
@@ -243,7 +245,7 @@ public class ModelGenerator {
 		return personList;
 	}
 	/**
-	 * @param fileName of the corresponding '.bookings' file in '/Flights/src/org/emoflon/flight/model/examples'
+	 * @param fileName of the corresponding '.bookings' file in '/Flights/src/org/emoflon/flight/model/definitions'
 	 * @param persons containing all persons named in the '.bookings' file
 	 * @param flightContainer containing all bookings named in the '.bookings' file
 	 * @return a booking container filled with bookings objects parsed from the '.bookings' file
@@ -255,7 +257,7 @@ public class ModelGenerator {
 		return bookings;
 	}
 	/**
-	 * @param fileName of the corresponding '.bookings' file in '/Flights/src/org/emoflon/flight/model/examples'
+	 * @param fileName of the corresponding '.bookings' file in '/Flights/src/org/emoflon/flight/model/definitions'
 	 * @param persons containing all persons named in the '.bookings' file
 	 * @param flightContainer containing all bookings named in the '.bookings' file
 	 * @return a booking list filled with bookings objects parsed from the '.bookings' file
