+++ src/org/emoflon/flight/model/util/ModelParser.java
@@ -7,13 +7,13 @@ import java.util.Scanner;
 
 public class ModelParser {
 	/**
-	 * @param fileName to parse in folder /Flights/src/org/emoflon/flight/model/examples
+	 * @param fileName to parse in folder /Flights/src/org/emoflon/flight/model/definitions
 	 * @return list of String arrays, containing each line of the file, that is not empty or begins with an '/',
 	 *  splitting each line at the ';' character
 	 */
 	public static ArrayList<String[]> parseFile(String fileName) {
 		try {
-			File myObj = new File("src\\org\\emoflon\\flight\\model\\examples\\" + fileName);
+			File myObj = new File("src\\org\\emoflon\\flight\\model\\definitions\\" + fileName);
 			ArrayList<String[]> out = new ArrayList<String[]>();
 			Scanner myReader = new Scanner(myObj);
 			while (myReader.hasNextLine()) {
@@ -33,7 +33,7 @@ public class ModelParser {
 	}
 	/**
 	 * @param index of the selected element in each split line
-	 * @param fileName to parse in folder /Flights/src/org/emoflon/flight/model/examples
+	 * @param fileName to parse in folder /Flights/src/org/emoflon/flight/model/definitions
 	 * @return string array filled with the selected element for each line
 	 */
 	public static String[] arrayParseFromFile(int index, String fileName) {
